Split code on dedicated file + Soft Reset
to optimize readability and make code more modular , code was split
among dedicated files, like Radio to manage radio comunication, log to
control log data, telemetry , display, menu do draw and manage display
menu, action to execute action upon menu selection

To adress a problem with SPIFlash memory where upon power flash was not
accessible, and a reset was always necessary, now at setup, there is a
detection if SPIFlash is accessible, and if not, a Soft Reset is
executed.
Apparently this is solving the problem, even if it seem to be an
hardware problem. - to be further tested/investigated.